+++ src/commands/rolebuttons/button.rs
@@ -46,7 +46,7 @@ async fn pressed(ctx: Context, mut interaction: MessageComponentInteraction) -> 
 
     let db = ctx.data.read().await.get::<DatabaseTypeMapKey>().unwrap().clone();
     let Some(server) =
-        RoleButtonServer::find().filter(role_button_server::Column::ServerId.eq(member.guild_id.0 as i64)).one(&db).await? else { return Err(anyhow!("Button pressed for an unregistered server.")) };
+        RoleButtonServer::find().filter(role_button_server::Column::ServerId.eq(member.guild_id.0)).one(&db).await? else { return Err(anyhow!("Button pressed for an unregistered server.")) };
 
     let role_id =
         interaction.data.custom_id.strip_prefix("role_").unwrap_or(&interaction.data.custom_id).parse::<RoleId>()?;
